back_end/requirements.txt
amqp==5.3.1
asgiref==3.8.1
babel==2.17.0
beautifulsoup4==4.13.3
billiard==4.2.1
celery==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.5
django-celery-beat==2.8.0
django-cors-headers==4.6.0
django-countries==7.6.1
django-filter==25.1
django-money==3.5.3
django-phonenumber-field==8.0.0
django-redis==5.4.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.9
idna==3.10
inflection==0.5.1
kombu==5.5.3
packaging==24.2
phonenumbers==8.13.54
prompt_toolkit==3.0.51
py-moneyed==3.0
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==6.0.0
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
